requirements.txt
django==4.2.10
selenium==4.18.1
pandas==2.2.1
numpy==1.26.3
plotly==5.18.0
beautifulsoup4==4.12.3
webdriver-manager==4.0.1
python-dotenv==1.0.1
requests==2.31.0
sentry-sdk==1.40.4
djangorestframework==3.14.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
gunicorn==21.2.0
apscheduler==3.10.4